mv site_label to avoid circular imports, use self.molecule instead of self.mol
